Check if file exists before removing it

+++ scripts/build.sh
@@ -1,5 +1,7 @@
-rm "build/pix_tester"
-clear
+if [-f "build/pix_tester"]; then
+	rm "build/pix_tester"
+fi
+
 cmake --build "build"
 hexdump -C -v "build/pix_tester" > "logs/hexdump.txt"
 objdump -D "build/pix_tester" > "logs/objdump.txt"
